requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
argon2-cffi==23.1.0
    # via bomb-insper (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==3.0.0
    # via icecream
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.1.8
    # via
    #   prisma
    #   uvicorn
colorama==0.4.6
    # via icecream
cryptography==44.0.2
    # via python-jose
ecdsa==0.19.1
    # via python-jose
executing==2.2.0
    # via icecream
fastapi==0.115.12
    # via bomb-insper (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   bomb-insper (pyproject.toml)
    #   prisma
icecream==2.1.4
    # via bomb-insper (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
jinja2==3.1.6
    # via prisma
markupsafe==3.0.2
    # via jinja2
msgpack==1.1.1
    # via bomb-insper (pyproject.toml)
nodeenv==1.9.1
    # via prisma
prisma==0.15.0
    # via bomb-insper (pyproject.toml)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   bomb-insper (pyproject.toml)
    #   fastapi
    #   prisma
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.8.1
    # via bomb-insper (pyproject.toml)
pygame==2.6.1
    # via bomb-insper (pyproject.toml)
pygments==2.19.1
    # via icecream
python-dotenv==1.1.0
    # via
    #   prisma
    #   pydantic-settings
python-jose==3.4.0
    # via bomb-insper (pyproject.toml)
python-multipart==0.0.20
    # via bomb-insper (pyproject.toml)
rsa==4.9
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
tomlkit==0.13.2
    # via prisma
typing-extensions==4.13.1
    # via
    #   fastapi
    #   prisma
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.0
    # via bomb-insper (pyproject.toml)
websockets==15.0.1
    # via bomb-insper (pyproject.toml)
